add RentalCart user musn finish active rental, then create new

+++ Areas/Customer/Controllers/RentalCartController.cs
@@ -6,6 +6,8 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Internal;
+using Microsoft.Extensions.Logging;
 using RentAMovies.Data;
 using RentAMovies.Models;
 using RentAMovies.Models.ViewModels;
@@ -58,49 +60,59 @@ namespace RentAMovies.Areas.Customer.Controllers
         [ValidateAntiForgeryToken]
         [ActionName("Index")]
         public async Task<IActionResult> SummaryPost()
-        {
+        {           
+
+
             var claimsIdentity = (ClaimsIdentity)User.Identity;
             var claim = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier);
 
+            var Status = await _context.RentalHeader.Where(c => c.UserId == claim.Value).ToListAsync();
 
-            detailCart.listCart2 = await _context.ShoppingCart.Where(c => c.ApplicationUserId == claim.Value).ToListAsync();
+            List<RentalHeader> FilterdStatus = Status.Where(p => p.Status.StartsWith("A")).ToList();
 
-
-            detailCart.RentalHeader.RentalDate = DateTime.Now;
-            detailCart.RentalHeader.UserId = claim.Value;
-            detailCart.RentalHeader.Status = SD.RentalprocesStatusAcitve;
-
-            List<RentalDetails> RentalDetailsList = new List<RentalDetails>();
-            _context.RentalHeader.Add(detailCart.RentalHeader);
-            await _context.SaveChangesAsync();
-
-
-
-            foreach (var item in detailCart.listCart2)
+            if (FilterdStatus.Count != 0)
             {
-                item.Movie = await _context.Movies.FirstOrDefaultAsync(m => m.Id == item.MovieId);
-                RentalDetails RentalDetails = new RentalDetails
-                {
-                    MovieId = item.MovieId,
-                    RentalHeaderId = detailCart.RentalHeader.Id,
-                    Status = true,
-            };
-                
-                _context.RentalDetails.Add(RentalDetails);
+                return RedirectToAction("Message", "RentalCart", new { id = detailCart.RentalHeader.Id });
 
             }
+            else
+            {
+                detailCart.listCart2 = await _context.ShoppingCart.Where(c => c.ApplicationUserId == claim.Value).ToListAsync();
 
-           
-            _context.ShoppingCart.RemoveRange(detailCart.listCart2);
-            HttpContext.Session.SetInt32(SD.ssShoppingCartCount, 0);
-            await _context.SaveChangesAsync();            
 
-           
-            //return RedirectToAction("Index", "Home");
-            return RedirectToAction("Confirm", "RentalOrder", new { id = detailCart.RentalHeader.Id });
+                detailCart.RentalHeader.RentalDate = DateTime.Now;
+                detailCart.RentalHeader.UserId = claim.Value;
+                detailCart.RentalHeader.Status = SD.RentalprocesStatusAcitve;
+
+                List<RentalDetails> RentalDetailsList = new List<RentalDetails>();
+                _context.RentalHeader.Add(detailCart.RentalHeader);
+                await _context.SaveChangesAsync();
+
+
+
+                foreach (var item in detailCart.listCart2)
+                {
+                    item.Movie = await _context.Movies.FirstOrDefaultAsync(m => m.Id == item.MovieId);
+                    RentalDetails RentalDetails = new RentalDetails
+                    {
+                        MovieId = item.MovieId,
+                        RentalHeaderId = detailCart.RentalHeader.Id,
+                        Status = true,
+                    };
+
+                    _context.RentalDetails.Add(RentalDetails);
+
+                }
+
+
+                _context.ShoppingCart.RemoveRange(detailCart.listCart2);
+                HttpContext.Session.SetInt32(SD.ssShoppingCartCount, 0);
+                await _context.SaveChangesAsync();
+                //return RedirectToAction("Index", "Home");
+                return RedirectToAction("Confirm", "RentalOrder", new { id = detailCart.RentalHeader.Id });
+            }
 
         }
-
         public async Task<IActionResult> Remove(int cartId)
         {
             var cart = await _context.ShoppingCart.FirstOrDefaultAsync(c => c.Id == cartId);
@@ -113,6 +125,11 @@ namespace RentAMovies.Areas.Customer.Controllers
 
 
             return RedirectToAction(nameof(Index));
-        }       
+        }
+
+        public IActionResult Message()
+        {
+            return View();
+        }
     }
 }
